+++ R/hagis-package.R
@@ -1,8 +1,8 @@
 
-#' Tools for Analysis of Plant Pathogen Pathotype Complexities, Distributions and Diversity 
+#' Tools for analysis of plant pathogen pathotype complexities, distributions and diversity 
 #'
 #' @name hagis
-#' @author Austin McCoy, mccoyaus@@msu.edu and Zachary Noel, noelzach@@msu.edu and
+#' @author Austin G. McCoy, mccoyaus@@msu.edu and Zachary Noel, noelzach@@msu.edu and
 #'  Adam H. Sparks, adam.sparks@@usq.edu.au and Martin Chilvers, chilvers@@msu.edu
 #' @docType package
 #'